app/requirements.txt
# Ensure modern build tools so sdist builds succeed
setuptools>=65.5.1
wheel>=0.40.0

altgraph>=0.17.3
certifi>=2022.9.24
charset-normalizer>=2.1.1
docopt>=0.6.2
et-xmlfile>=1.1.0
future>=0.18.2
idna>=3.4
numpy>=1.26,<2.0
openpyxl>=3.1.2
pandas>=2.1,<3.0
pefile>=2022.5.30
pyinstaller>=5.6.2
pyinstaller-hooks-contrib>=2022.13
python-dateutil>=2.8.2
pytz>=2022.5
pywin32-ctypes>=0.2.0
PyYAML>=6.0
requests>=2.28.1
six>=1.16.0
urllib3>=1.26.12
